+++ train_sdf/test.py
@@ -50,7 +50,6 @@ def test_modulations():
             # point_cloud = data # filename = path to the csv file of sdf data
             point_cloud, cls_name, mesh_name = data # filename = path to the csv file of sdf data
             # filename = filename[0] # filename is a tuple
-
             # cls_name = filename.split("/")[-3]
             # mesh_name = filename.split("/")[-2]
             outdir = os.path.join(recon_dir, "{}/{}".format(cls_name, mesh_name))
@@ -59,7 +58,7 @@ def test_modulations():
             
             # given point cloud, create modulations (e.g. 1D latent vectors)
             print("point_cloud.shape: ", point_cloud.shape, point_cloud.dtype)
-            plane_features = model.sdf_model.pointnet.get_plane_features(point_cloud.cuda())  # tuple, 3 items with ([1, D, resolution, resolution])
+            plane_features = model.sdf_model.pointnet.get_plane_features(point_cloud.float().cuda())  # tuple, 3 items with ([1, D, resolution, resolution])
             plane_features = torch.cat(plane_features, dim=1) # ([1, D*3, resolution, resolution])
             recon = model.vae_model.generate(plane_features) # ([1, D*3, resolution, resolution])
             #print("mesh filename: ", mesh_filename)
